refactor: extract EPUB semantic text rendering
